+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Cast 
+
+#### Added
+
+- Aliases in `snfoundry.toml`. Read more [here](https://foundry-rs.github.io/starknet-foundry/starknet/aliases.html).
+- `--contract-address @alias` syntax in `sncast call` and `sncast invoke`.
+
 ## [0.61.0] - 2026-05-26
 
 ### Forge